newest

+++ RogueMultiplayer/src/Packet/PacketUpdatePlayerSprite.java
@@ -3,5 +3,5 @@ package Packet;
 public class PacketUpdatePlayerSprite {
 	
 	public int id;
-	public int spriteX, spriteY;
+	public byte spriteX, spriteY;
 }
